# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Muslim World API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Site {
  id: number;
  /** English name of the site */
  name: string;
  /** Arabic name of the site */
  arabicName: string;
  /** Geographic region (e.g. Arabian Peninsula, North Africa, South Asia) */
  region: string;
  country: string;
  latitude: number;
  longitude: number;
  /** mosque, shrine, palace, fortress, city */
  category: string;
  shortDescription: string;
  /** @nullable */
  yearFounded: string | null;
  /** Brief note on Islamic historical significance */
  significance: string;
  isFeatured?: boolean;
  /** @nullable */
  imageUrl?: string | null;
  /**
     * Path to a GLTF/GLB 3D model file for this site
     * @nullable
     */
  modelUrl?: string | null;
  /**
     * Approximate worshipper capacity
     * @nullable
     */
  capacity?: number | null;
  /**
     * Total site area in square metres
     * @nullable
     */
  areaSqm?: number | null;
  /**
     * Secondary function, e.g. "Mosque & University"
     * @nullable
     */
  dualUse?: string | null;
  /** Whether Eid prayer is held here */
  eidPrayer?: boolean;
  /** Whether this is a notable Ramadan visit destination */
  ramadanVisit?: boolean;
  /** Whether Jumu'ah (Friday) prayer is held here */
  jumaPrayer?: boolean;
}

export interface Hotspot {
  id: number;
  siteId: number;
  /** Name of the architectural feature (e.g. Minaret, Green Dome) */
  label: string;
  /** Detailed description shown when user clicks the hotspot */
  description: string;
  /** X position on the 3D model (normalized -1 to 1) */
  positionX: number;
  /** Y position on the 3D model (normalized -1 to 1) */
  positionY: number;
  /** Z position on the 3D model (normalized -1 to 1) */
  positionZ: number;
  /**
     * Arabic term for this feature
     * @nullable
     */
  arabicTerm?: string | null;
  /** @nullable */
  historicalPeriod?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface SiteDetail {
  id: number;
  name: string;
  arabicName: string;
  region: string;
  country: string;
  latitude: number;
  longitude: number;
  category: string;
  shortDescription: string;
  fullDescription: string;
  /** @nullable */
  yearFounded: string | null;
  significance: string;
  isFeatured?: boolean;
  /** @nullable */
  imageUrl?: string | null;
  /**
     * Path to a GLTF/GLB 3D model file for this site
     * @nullable
     */
  modelUrl?: string | null;
  /** @nullable */
  architecturalStyle?: string | null;
  /**
     * Approximate worshipper capacity
     * @nullable
     */
  capacity?: number | null;
  /**
     * Total site area in square metres
     * @nullable
     */
  areaSqm?: number | null;
  /**
     * Secondary function, e.g. "Mosque & University"
     * @nullable
     */
  dualUse?: string | null;
  eidPrayer?: boolean;
  ramadanVisit?: boolean;
  jumaPrayer?: boolean;
  hotspots: Hotspot[];
}

export interface UpdateSite {
  name?: string;
  arabicName?: string;
  region?: string;
  country?: string;
  latitude?: number;
  longitude?: number;
  category?: string;
  shortDescription?: string;
  fullDescription?: string;
  /** @nullable */
  yearFounded?: string | null;
  significance?: string;
  isFeatured?: boolean;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  modelUrl?: string | null;
  /** @nullable */
  architecturalStyle?: string | null;
  /** @nullable */
  capacity?: number | null;
  /** @nullable */
  areaSqm?: number | null;
  /** @nullable */
  dualUse?: string | null;
  eidPrayer?: boolean;
  ramadanVisit?: boolean;
  jumaPrayer?: boolean;
}

export interface RegionGroup {
  region: string;
  count: number;
  sites: Site[];
}