# lib/api-zod/src/generated/types/site.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Muslim World API
 * OpenAPI spec version: 0.1.0
 */

export interface Site {
  id: number;
  /** English name of the site */
  name: string;
  /** Arabic name of the site */
  arabicName: string;
  /** Geographic region (e.g. Arabian Peninsula, North Africa, South Asia) */
  region: string;
  country: string;
  latitude: number;
  longitude: number;
  /** mosque, shrine, palace, fortress, city */
  category: string;
  shortDescription: string;
  /** @nullable */
  yearFounded: string | null;
  /** Brief note on Islamic historical significance */
  significance: string;
  isFeatured?: boolean;
  /** @nullable */
  imageUrl?: string | null;
  /**
     * Path to a GLTF/GLB 3D model file for this site
     * @nullable
     */
  modelUrl?: string | null;
  /**
     * Approximate worshipper capacity
     * @nullable
     */
  capacity?: number | null;
  /**
     * Total site area in square metres
     * @nullable
     */
  areaSqm?: number | null;
  /**
     * Secondary function, e.g. "Mosque & University"
     * @nullable
     */
  dualUse?: string | null;
  /** Whether Eid prayer is held here */
  eidPrayer?: boolean;
  /** Whether this is a notable Ramadan visit destination */
  ramadanVisit?: boolean;
  /** Whether Jumu'ah (Friday) prayer is held here */
  jumaPrayer?: boolean;
}