# lib/api-zod/src/generated/types/siteDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Muslim World API
 * OpenAPI spec version: 0.1.0
 */
import type { Hotspot } from './hotspot';

export interface SiteDetail {
  id: number;
  name: string;
  arabicName: string;
  region: string;
  country: string;
  latitude: number;
  longitude: number;
  category: string;
  shortDescription: string;
  fullDescription: string;
  /** @nullable */
  yearFounded: string | null;
  significance: string;
  isFeatured?: boolean;
  /** @nullable */
  imageUrl?: string | null;
  /**
     * Path to a GLTF/GLB 3D model file for this site
     * @nullable
     */
  modelUrl?: string | null;
  /** @nullable */
  architecturalStyle?: string | null;
  /**
     * Approximate worshipper capacity
     * @nullable
     */
  capacity?: number | null;
  /**
     * Total site area in square metres
     * @nullable
     */
  areaSqm?: number | null;
  /**
     * Secondary function, e.g. "Mosque & University"
     * @nullable
     */
  dualUse?: string | null;
  eidPrayer?: boolean;
  ramadanVisit?: boolean;
  jumaPrayer?: boolean;
  hotspots: Hotspot[];
}