# lib/api-zod/src/generated/types/updateSite.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * The Muslim World API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSite {
  name?: string;
  arabicName?: string;
  region?: string;
  country?: string;
  latitude?: number;
  longitude?: number;
  category?: string;
  shortDescription?: string;
  fullDescription?: string;
  /** @nullable */
  yearFounded?: string | null;
  significance?: string;
  isFeatured?: boolean;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  modelUrl?: string | null;
  /** @nullable */
  architecturalStyle?: string | null;
  /** @nullable */
  capacity?: number | null;
  /** @nullable */
  areaSqm?: number | null;
  /** @nullable */
  dualUse?: string | null;
  eidPrayer?: boolean;
  ramadanVisit?: boolean;
  jumaPrayer?: boolean;
}